heartrate_processor/requirements.txt
Flask>=2.2.3
paho-mqtt>=1.6.1
pandas>=2.0.0